fix warning

+++ libs/hwdrivers/src/CIMUXSens_MT4.cpp
@@ -375,7 +375,7 @@ void CIMUXSens_MT4::initialize()
 					"CIMUXSens_MT4: No 'portname' was specified and no "
 					"compatible XSens device was found in the system (%u "
 					"devices connected)",
-					portInfoArray.size());
+					static_cast<unsigned int>(portInfoArray.size()));
 
 			if (m_verbose)
 				cout << "[CIMUXSens_MT4] Found " << portInfoArray.size()
